Add wind and solar template filling
