+++ rules/wind-and-solar.smk
@@ -38,6 +38,20 @@ rule download_capacity_factors_wind_and_solar:
     shell: "curl -sLo {output} '{params.url}'"
 
 
+rule area_to_capacity_limits:
+    message: "Use technology densities to convert wind & solar {wildcards.resolution} available area to capacity limits."
+    input:
+        script = script_dir + "wind-and-solar/capacity_limits.py",
+        units = rules.units_without_shape.output[0],
+        land_eligibility_km2 = rules.potentials.output.land_eligibility_km2,
+    params:
+        max_power_density = config["parameters"]["maximum-installable-power-density"],
+        roof_shares = config["parameters"]["roof-share"],
+    output: "build/data/{resolution}/wind-and-solar-capacity-limits.csv"
+    conda: "../envs/default.yaml"
+    script: "../scripts/wind-and-solar/capacity_limits.py"
+
+
 rule capacity_factors_onshore_wind_and_solar:
     message: "Generate capacityfactor time series disaggregated by location on "
              "{wildcards.resolution} resolution for {wildcards.technology}."
@@ -54,7 +68,7 @@ rule capacity_factors_onshore_wind_and_solar:
         trim_ts = config["capacity-factors"]["trim-ninja-timeseries"]
     wildcard_constraints:
         technology = "wind-onshore|rooftop-pv|open-field-pv|rooftop-pv-n|rooftop-pv-e-w|rooftop-pv-s-flat"
-    output: "build/model/{resolution}/capacityfactors-{technology}.csv"
+    output: "build/model/timeseries_data/{resolution}/supply/capacityfactors-{technology}.csv"
     conda: "../envs/geo.yaml"
     script: "../scripts/capacityfactors.py"
 
@@ -73,6 +87,6 @@ rule capacity_factors_offshore:
         first_year = config["scope"]["temporal"]["first-year"],
         final_year = config["scope"]["temporal"]["final-year"],
         trim_ts = config["capacity-factors"]["trim-ninja-timeseries"]
-    output: "build/model/{resolution}/capacityfactors-wind-offshore.csv"
+    output: "build/model/timeseries_data/{resolution}/supply/capacityfactors-wind-offshore.csv"
     conda: "../envs/geo.yaml"
     script: "../scripts/capacityfactors_offshore.py"
